Source: Change types of values to reflect their limits
bench: 4321924

+++ Source/evaluate.h
@@ -3,5 +3,5 @@
 
 #include "structs.h"
 
-int evaluate(position_t *pos, accumulator_t *accumulator);
+int16_t evaluate(position_t *pos, accumulator_t *accumulator);
 #endif
